% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_roc.R
\name{average_rocdata}
\alias{average_rocdata}
\title{average_rocdata}
\usage{
average_rocdata(rocdata, n = 1000)
}
\arguments{
\item{rocdata}{a \code{\link{extract_rocdata}} object.}

\item{n}{the number of thresholds in \[0,1\] to evaluate.}
}
\value{
a data.frame object of class "rocdata" which can be plotted.
Contents:
\itemize{
\item{\code{Sens} : Sensitivity}
\item{\code{InvSpec} : 1 - Specificity}
\item{\code{alpha} : threshold}
\item{\code{run} : a label indicating this is averaged}
}
}
\description{
Function calculates the average of a set of logistic regression ROC curves.
For a range of thresholds between 0 and 1 the sensitivity and
specificity are extracted and mean averaged over the set of curves.
This is intended to be used with k-fold cross-validation data.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Warning: The averaging of roc-curves is difficult
    (e.g. \\href\{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4112395/\}\{
    Chen & Samuelson Br J. Radiol Aug 2014\})
    , particularly while preserving the AUROC such that the AUROC
    for the average ROC-curve is equal to the average AUROC of the
    component ROC-curves.

    This function is for a rough display of average performance
    over bootstraps or repeats of k-fold cross-validation.

    Minimal testing suggests that there should be agreement in the
    AUROCs to the third decimal place.
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link{plot.rocdata}}, \code{\link{extract_rocdata}}
}
