% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_calibration.R
\name{calibration}
\alias{calibration}
\alias{calibration.default}
\title{calibration}
\usage{
calibration(x, ...)

\method{calibration}{default}(x, ppid = "prediction", ooid = "reference", gid = "label", ...)
}
\arguments{
\item{x}{an object containing predicted probabilities and observed outcomes,
from which calibration can be extracted.}

\item{...}{arguments to pass on}

\item{ppid}{indicator for predicted probability element in x
(column "name" or index)}

\item{ooid}{indicator for observed outcome element in x
(column "name" or index)}

\item{gid}{indicator for grouping variable in x
(column "name" or index). Set to NA to force no grouping.}
}
\value{
calibration intercept and calibration slope
}
\description{
calculates 'weak' (i.e. intercept + slope) calibration equivalent to the
values returned by the val.prob function in the rms package
(which accompanies Frank Harrell's Regression Modelling Strategies book).

Default function behaviour assumes input is a
list/data.frame with required vectors as elements.
}
\details{
Calibration is not returned via \code{\link{performance}} because of the
computational overhead of model fitting.
}
