% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_classperformance.R
\name{casesummary.performance}
\alias{casesummary.performance}
\title{casesummary.performance}
\usage{
casesummary.performance(object, type = c("both", "data", "summary"), ...)
}
\arguments{
\item{object}{a \code{\link{performance}} object?}

\item{type}{What should be returned?
\itemize{
\item{\code{data} - The true and estimated classifications.}
\item{\code{summary} - The mean proportion correct}
}}

\item{...}{additional arguments (not currently used)}
}
\description{
What proportion of the time were subjects correctly classified in a
\code{\link{performance}} object.
}
