% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_innerloop.R
\name{coef.multialpha.repeated.cv.glmnet}
\alias{coef.multialpha.repeated.cv.glmnet}
\title{coef.multialpha.repeated.cv.glmnet}
\usage{
\method{coef}{multialpha.repeated.cv.glmnet}(object, newx = NULL,
  alpha = NULL, s = NULL, ...)
}
\arguments{
\item{object}{a a \code{\link{multialpha.repeated.cv.glmnet}} object.}

\item{newx}{matrix of new values for x at which predictions are required.
Note: no impact when type is "coefficients", "class" or "nonzero".}

\item{alpha}{the penalty type alpha at which prediction is required.
Leave NULL to use the cv-optimised value.}

\item{s}{The penalty amount lambda at which prediction is required.
Leave NULL to use the cv-optimised value.}

\item{...}{passed to \code{\link[glmnet]{predict.glmnet}}}
}
\description{
obtain coeffiicents from a
\code{\link{multialpha.repeated.cv.glmnet}} object.
This is a wrapper for \code{\link{predict.multialpha.repeated.cv.glmnet}}
}
\seealso{
\code{\link{predict.multialpha.repeated.cv.glmnet}}
\code{\link[glmnet]{predict.cv.glmnet}},
\code{\link[glmnet]{predict.glmnet}}
}
