% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_referencemodels.R
\name{coef_reflogreg}
\alias{coef_reflogreg}
\title{coef_reflogreg}
\usage{
coef_reflogreg(refobj)
}
\arguments{
\item{refobj}{a set of reference models provided by
\code{\link{reflogreg}}}
}
\value{
a data.frame with columns for:
\itemize{
\item{\code{glm} - unless PCA was performed}
\item{\code{univariate} - unless univariate was not requested}
}
}
\description{
Returns model coefficients (betas) for glm/univariate reference models.
}
