% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{cv.glmnet.modelsummary}
\alias{cv.glmnet.modelsummary}
\title{cv.glmnet.modelsummary}
\usage{
cv.glmnet.modelsummary(mod, alpha = NA, rep = NA)
}
\arguments{
\item{mod}{a cv.glmnet object}

\item{alpha}{an optional label}

\item{rep}{an optional label}
}
\description{
return a dataframe of cv.glmnet results.
}
