% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{cv_classperformance_glm}
\alias{cv_classperformance_glm}
\title{cv_classperformance_glm}
\usage{
cv_classperformance_glm(y, data, f = "~.", folds = NULL, k = 10,
  nrep = 2, family = "binomial", opt.ystratify = TRUE,
  opt.uniquefolds = FALSE, ...)
}
\arguments{
\item{y}{outcome vector (numeric or factor)}

\item{data}{predictors in a data.frame}

\item{f}{a formula to apply to x}

\item{folds}{This is a list where each element is an integer vector
of length \emph{n_cases}. The integer for each case labels it as belonging
to a fold \emph{1:n_folds}. This argument implicitly sets the number of repeats
and the k in repeated k-fold cv.}

\item{k}{the number of folds for k-fold cross-validation.}

\item{nrep}{the number of repetitions}

\item{family}{Response type (see above)}

\item{opt.ystratify}{Boolean.
Outer and inner sampling is stratified by outcome.
This is implemented with \code{\link[caret]{createFolds}}}

\item{opt.uniquefolds}{Boolean.
In most circumstances folds will be unique. This requests
that random folds are checked for uniqueness in inner and outer loops.
Currently it warns if non-unique values are found.}

\item{...}{other arguments}
}
\value{
A list containing the following:
\itemize{
\item{glm.performance - summary(classperformance(x))
for the uncrossvalidated model}
\item{cv.performance - report_classperformance_summary(cv.fits)
for the crossvalidated model}
\item{folds - the folds used in cross-validation}
\item{call - the function call}
}
}
\description{
Cross-validated estimates of model performance by
repeated k-fold cross-validation.
}
\details{
This function is nothing revolutionary. The idea is to
extend \code{\link{boot}{cv.glm}} with an interface that better matches
the other functions in this package.

The additions are:
\itemize{
\item{Repeated k-fold rather than single k-fold}
\item{Option to provide the fold membership}
\item{Default use of stratified sampling by outcome class}
\item{Performance assessed with \code{\link{summary.classperformance}}}
}
}
\seealso{
\code{\link[boot]{cv.glm}}, \code{\link{classperformance}}
}
