% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{cv_performance_glm}
\alias{cv_performance_glm}
\title{cv_performance_glm}
\usage{
cv_performance_glm(
  y,
  data,
  f = "~.",
  folds = NULL,
  k = 10,
  nrep = 2,
  family = "binomial",
  opt.ystratify = TRUE,
  opt.uniquefolds = FALSE,
  return_summary = TRUE,
  ...
)
}
\arguments{
\item{y}{outcome vector (numeric or factor)}

\item{data}{predictors in a data.frame}

\item{f}{a formula to apply to x}

\item{folds}{This is a list where each element is an integer vector
of length \emph{n_cases}. The integer for each case labels it as belonging
to a fold \emph{1:n_folds}. This argument overrides the number of repeats
and the k in repeated k-fold cv.}

\item{k}{the number of folds for k-fold cross-validation.}

\item{nrep}{the number of repetitions}

\item{family}{Response type (see above). Either a character string representing
one of the built-in families, or else a \code{glm()} family object.}

\item{opt.ystratify}{Boolean.
Outer and inner sampling is stratified by outcome.
This is implemented with \code{\link[caret]{createFolds}}}

\item{opt.uniquefolds}{Boolean.
In most circumstances folds will be unique. This requests
that random folds are checked for uniqueness in inner and outer loops.
Currently it warns if non-unique values are found.}

\item{return_summary}{bool. return summarised performance (default), or
\code{\link{performance}} objects for further analysis (set to FALSE)}

\item{...}{other arguments}
}
\value{
A list containing the following:
\itemize{
\item{glm.performance - summary(performance(x))
for the uncrossvalidated model}
\item{cv.performance - report_performance_summary(cv.fits)
for the crossvalidated model}
\item{folds - the folds used in cross-validation}
\item{call - the function call}
}
}
\description{
Cross-validated estimates of model performance by
repeated k-fold cross-validation.
}
\details{
This function is nothing revolutionary. The idea is to
extend \code{\link{boot}{cv.glm}} with an interface that better matches
the other functions in this package.

The additions are:
\itemize{
\item{Repeated k-fold rather than single k-fold}
\item{Option to provide the fold membership}
\item{Default use of stratified sampling by outcome class}
\item{Performance assessed with \code{\link{summary.performance}}}
}
}
\seealso{
\code{\link[boot]{cv.glm}}, \code{\link{performance}}
}
