% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet.R
\docType{package}
\name{dCVnet-package}
\alias{dCVnet-package}
\title{dCVnet: Doubly Cross-Validated Elastic-Net Regularised (Generalised) Linear Models}
\description{
dCVnet runs binary logistic regression with elastic-net regulatisation.
Optimal hyperparameters are selected by double (nested) cross-validation with
model performance evaluated in an independent outer loop.
}
\details{
The lambda and alpha hyperparameters of the elastic-net permit models ranging
from effectively unregularised to heavily-regularised models with a mixture
of two types of regularisation. The amount and type of regularisation will be
selected on the basis of what generalises best in the inner cross-validation.
L2 (ridge) and L1 (LASSO) effectively produce dimensionality reduction and
variable selection in the predictors.

The fully nested cross-validation keeps the final performance estimates
'honest' by reducing optimism bias.

dCVnet is a wrapper around \code{glmnet::cv.lognet}.
}
