% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_referencemodels.R
\name{dCVnet_refmodels}
\alias{dCVnet_refmodels}
\title{dCVnet_refmodels}
\usage{
dCVnet_refmodels(object)
}
\arguments{
\item{object}{a dCVnet object}
}
\value{
a list containing: \itemize{
\item{\code{glm} - the multiple-predictor model (possibly PCA-reduced)}
\item{\code{univariate} - a \code{glmlist} of models, one for each predictor}
}
}
\description{
calculate some reference models to help interpret dCVnet performance.
models for \eqn{n} observations of \eqn{p} predictors calculated are:
\itemize{
\item{a logistic regression using all variables (if \eqn{n > 5 * p}),
otherwise a logistic regression on the first \eqn{round(n / 5)}
principal compoments}
\item{a series of logistic regressions, one for each column in the
design matrix - a mass univariate approach}
}
}
\details{
The univariate component has a class ('glmlist') used in some summary
functions. This is not currently correctly implemented.
}
