% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{extract_glmnet_alpha}
\alias{extract_glmnet_alpha}
\title{extract_glmnet_alpha}
\usage{
extract_glmnet_alpha(mod)
}
\arguments{
\item{mod}{a \code{\link[glmnet]{glmnet}} or \code{\link[glmnet]{cv.glmnet}}
model}
}
\description{
glmnet model objects do not explicitly include the alpha value which was
used when the model was fit. This utility function tries to extract the
value of alpha used to fit a glmnet object - this is not always
possible (e.g. if a named variable in another script is used to set alpha
in the call to glmnet)
}
\details{
Note: dCVnet will quietly extend the glmnet models it fits to include
additional information. This information includes the alpha value
(see \code{\link{set_glmnet_alpha}}). This dCVnet specified
alpha value will be extracted with priority over alpha values implied
in the model without checking.
The dCVnet extension of the (cv.)glmnet S3 class is very basic.
}
\seealso{
\code{\link{set_glmnet_alpha}}
}
