% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_roc.R
\name{extract_rocdata}
\alias{extract_rocdata}
\title{extract_rocdata}
\usage{
extract_rocdata(performance, invertprob = FALSE)
}
\arguments{
\item{performance}{a \code{\link{performance}} object.}

\item{invertprob}{boolean. Should class probabilities be inverted?
If the ROC curve appears under the diagonal then toggle this option.}
}
\value{
a data.frame object of class "rocdata" which can be plotted.
Contents:
\itemize{
\item{\code{Sens} : Sensitivity}
\item{\code{InvSpec} : 1 - Specificity}
\item{\code{alpha} : threshold}
\item{\code{run} : label from \code{\link{performance}}}
}
}
\description{
Function reads ROC data from a \code{\link{performance}} object.
Sensitivity and specificity are calculated
at incrementing class probability thresholds.
}
\details{
Uses \pkg{ROCR} functions: \code{\link[ROCR]{prediction}}, and
\code{\link[ROCR]{performance}}
}
\seealso{
\code{\link{plot.rocdata}}
}
