% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_innerloop.R, R/dCVnet_main.R,
%   R/dCVnet_performance.R
\name{family.multialpha.repeated.cv.glmnet}
\alias{family.multialpha.repeated.cv.glmnet}
\alias{family.dCVnet}
\alias{family.performance}
\title{family.dCVnet}
\usage{
\method{family}{multialpha.repeated.cv.glmnet}(object, ...)

\method{family}{dCVnet}(object, ...)

\method{family}{performance}(object, ...)
}
\arguments{
\item{object}{a a \code{\link{dCVnet}} object.}

\item{...}{further arguments passed to methods}
}
\value{
character vector (of length 1)
}
\description{
Extracts the model family from a dCVnet object as a string
}
\section{Functions}{
\itemize{
\item \code{family.multialpha.repeated.cv.glmnet}: family for multialpha objects

\item \code{family.performance}: family for model performance objects
}}

