% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_performance.R
\name{InternalPerformanceSummaryFunctions}
\alias{InternalPerformanceSummaryFunctions}
\alias{perf_nomial}
\alias{perf_cont}
\alias{perf_cox}
\alias{perf_mgaussian}
\title{Internal summary functions for performance objects from
different model families.}
\usage{
perf_nomial(object, short = FALSE, somersD = FALSE, pvprevalence = "observed")

perf_cont(object, short = FALSE, somersD = FALSE, pvprevalence = "observed")

perf_cox(object, short = FALSE, somersD = FALSE, pvprevalence = "observed")

perf_mgaussian(
  object,
  short = FALSE,
  somersD = FALSE,
  pvprevalence = "observed"
)
}
\arguments{
\item{object}{a \code{\link{performance}} object.}

\item{short}{(bool) return a core set of performance measures.}

\item{somersD}{(bool) calculate the computationally expensive Somers' D for
certain families (gaussian, poisson)}

\item{pvprevalence}{argument for adjustment of PPV/NPV calculation for
binomial or multinomial families. Either "observed" to use the observed
prevalence, or a number \code{[0,1]} (for binomial),
or a vector of length n_categories (for multinomial) containing numerics
in \code{[0,1]}.}
}
\description{
For \strong{binomial and multinomial} family models:
Various performance metrics from \code{\link[caret]{confusionMatrix}};
AUC (or mAUC) (\code{\link[ModelMetrics]{auc}});
Brier Score (\code{\link[ModelMetrics]{brier}})

For \strong{Gaussian/Poisson} family models:
Root Mean Square Error (RMSE; \code{\link[ModelMetrics]{rmse}});
Mean Absolute Error (MAE; \code{\link[ModelMetrics]{mae}});
Pearson's correlation r and r^2 between observed and predicted (R, R2);
simple linear calibration intercept and slope (cal_Intercept, cal_Slope).
Brier score, which in this context is the Mean Square Error == RMSE^2;
RMSE and MSE scaled by the S.D. of y (SDScaledRMSE, SDScaledMAE);
Optionally Somer's Dxy can be returned (SomersDxy;
\code{\link[DescTools]{SomersDelta}}).
For the Poisson family performance measures above but calculated
from \code{\link{log1p}} transformed data are also returned.

For \strong{Cox} family models:
Harrell's C-index and Dxy using the \code{\link[Hmisc]{rcorr.cens}}
function.

For \strong{multivariate Gaussian} family models:
Runs \link{perf_cont} for each outcome.
}
\section{Functions}{
\itemize{
\item \code{perf_nomial()}: Used for binomial and multinomial families

\item \code{perf_cont()}: Used for gaussian and poisson families

\item \code{perf_cox()}: Used for cox models

\item \code{perf_mgaussian()}: Used for multivariate gaussian models

}}
