% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_innerloop.R
\name{is_multialpha.repeated.cv.glmnet}
\alias{is_multialpha.repeated.cv.glmnet}
\title{is_multialpha.repeated.cv.glmnet}
\usage{
is_multialpha.repeated.cv.glmnet(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE if object appears to be a valid instance of
multialpha.repeated.cv.glmnet, FALSE otherwise.
}
\description{
Checks if x is a mutlialpha.repeated.cv.glmnet object.
}
