% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{lambda_rangefinder}
\alias{lambda_rangefinder}
\title{lambda_rangefinder}
\usage{
lambda_rangefinder(y, x, alphalist, prop, niter = 10000)
}
\arguments{
\item{y}{binary outcome (from \code{\link{parse_dCVnet_input}})}

\item{x}{predictor matrix (from \code{\link{parse_dCVnet_input}})}

\item{alphalist}{alphas to consider, e.g. \code{c(0.2, 0.5, 0.8)}}

\item{prop}{proportion of the input to bootstrap. This should be the
fraction of observations remaining in the inner loop CV.
for example: if outer and inner loops are both 5-fold,
prop should be 0.8 * 0.8 = 0.64}

\item{niter}{the number of times to bootstrap.}
}
\value{
\code{maxlambda} - the largest bootstrapped lambda value
for each alpha requested.
}
\description{
What range of lambda penalty amounts should we consider for a dataset?
The maximum for a given dataset is determined by a simple formula,
but this can vary from subset to subset. This code bootstraps
the calculation of max-lambda for subsets of the input data.
}
