% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{lambda_seq_list}
\alias{lambda_seq_list}
\title{lambda_seq_list}
\usage{
lambda_seq_list(maxlambdas, nlambda, lambda_min_ratio)
}
\arguments{
\item{maxlambdas}{max lambdas vector,
output by \code{\link{lambda_rangefinder}}}

\item{nlambda}{how many steps to take}

\item{lambda_min_ratio}{what fraction of the maximum to take as the minimum.
no default is set, glmnet recommends
\code{ifelse(nobs<nvars,0.01,0.0001)}.}
}
\description{
produce descending lambda sequences given a maximum, minimum fraction and
the number of steps
}
