% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_logger.R
\name{log_results_to_excel}
\alias{log_results_to_excel}
\title{log_results_to_excel}
\usage{
log_results_to_excel(object, file, referencemodel = TRUE)
}
\arguments{
\item{object}{a \code{\link{dCVnet}} object.}

\item{file}{path to a writable excel file (i.e. must not be open/locked).}

\item{referencemodel}{either set to \code{TRUE} to calculate
reference models, or pass a pre-calculated \code{\link{reflogreg}}
object.
Any other values (e.g. \code{FALSE}) suppress the
reference model calcuations.}
}
\value{
No return value.
}
\description{
write the results of a \code{\link{dCVnet}} model to a multi-worksheet
xlsx format excel file.
}
