% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modsel_perf}
\alias{modsel_perf}
\title{Used for vignette: dCVnet-limitations}
\format{
An object of class \code{list} of length 2.
}
\usage{
modsel_perf
}
\description{
Used for vignette: dCVnet-limitations
}
\keyword{datasets}
