% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{parse_alphalist}
\alias{parse_alphalist}
\title{parse_alphalist}
\usage{
parse_alphalist(alphalist)
}
\arguments{
\item{alphalist}{a numeric of alpha values.}
}
\value{
non-duplicated, non-missing alpha values between (0, 1] -
i.e. excluding zero.
}
\description{
Check and process a numeric of alpha values.
}
\details{
\itemize{
\item{Zero alphas (i.e. pure L2 / ridge penalty) do not currently work
due to a bug in glmnet. Zeros are replaced with a small non-zero value.
For a 'pure' L2/ridge regression use a different package.}
\item{it is often the case that small alphas (<0.1) are slow to fit.}
}
}
