% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{parseddata_summary}
\alias{parseddata_summary}
\title{parseddata_summary}
\usage{
parseddata_summary(object)
}
\arguments{
\item{object}{either a \code{\link{dCVnet}} object, or a parsed dataset
output from \code{\link{parse_dCVnet_input}}}
}
\value{
a list containing: \itemize{
\item{ \code{OutcomeData} - frequencies, proportions and level names
for the outcome factor}
\item{ \code{PredictorData} - descriptives for the predictor matrix}
}
}
\description{
Simple descriptives for dCVnet's input (parsed dataset).
}
\details{
This acts on a model matrix, so factor variables will be indicator coded.
nnz should be informative for such variables.
}
