% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{plot.dCVnet}
\alias{plot.dCVnet}
\title{plot.dCVnet}
\usage{
\method{plot}{dCVnet}(x, type = "tuning", ...)
}
\arguments{
\item{x}{a \code{\link{dCVnet}} object}

\item{type}{one of
\itemize{
\item{\code{"tuning"} - hyperparameter tuning plots
(\code{\link{tuning_plot_dCVnet}})}
\item{\code{"roc"} - a roc plot (\code{\link{plot.rocdata}})}
}}

\item{...}{additional arguments passed to plot functions above.}
}
\description{
Produces either a plot of the tuning parameters and inner loop performance
per fold of the outer loop, or the final roc plots of the outer loop.
\itemize{
\item{\code{tuning} - selected tuning parameters over repeated folds of the
outerloop}
\item{\code{roc} - Sensitivity vs. (1 - Specificity) plot.}
}
}
