% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{predict.dCVnet}
\alias{predict.dCVnet}
\title{predict.dCVnet}
\usage{
\method{predict}{dCVnet}(object, newx, ...)
}
\arguments{
\item{object}{a a \code{\link{dCVnet}} object.}

\item{newx}{matrix of new values for x at which predictions are required.
Note: no impact when type is "coefficients", "class" or "nonzero".}

\item{...}{passed to \code{\link{predict.multialpha.repeated.cv.glmnet}},
then \code{\link[glmnet]{predict.glmnet}}}
}
\value{
predictions of the requested type.
}
\description{
predict method for a \code{\link{dCVnet}} object.
predictions come from the "production" model fitted to the full data.
As a result they do not reflect cross-validated/out-of-sample
performance.
}
