% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_innerloop.R
\name{predict.multialpha.repeated.cv.glmnet}
\alias{predict.multialpha.repeated.cv.glmnet}
\title{predict.multialpha.repeated.cv.glmnet}
\usage{
\method{predict}{multialpha.repeated.cv.glmnet}(object, alpha = NULL,
  s = NULL, ...)
}
\arguments{
\item{object}{a a \code{\link{multialpha.repeated.cv.glmnet}} object.}

\item{alpha}{the penalty type alpha at which prediction is required.
Leave NULL to use the cv-optimised value.}

\item{s}{The penalty amount lambda at which prediction is required.
Leave NULL to use the cv-optimised value.}

\item{...}{passed to \code{\link[glmnet]{predict.glmnet}}}
}
\description{
obtain predictions from a
\code{\link{multialpha.repeated.cv.glmnet}} object.
Uses the 'best' lambda (lambda.min or lambda.1se) determined by
type.lambda (unless requested otherwise).
}
\seealso{
\code{\link[glmnet]{predict.cv.glmnet}},
\code{\link[glmnet]{predict.glmnet}}
}
