% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{prediction_error_plot}
\alias{prediction_error_plot}
\title{prediction_error_plot}
\usage{
prediction_error_plot(x, bylabel = FALSE, log = FALSE, jitter = 0, plot = TRUE)
}
\arguments{
\item{x}{a \code{\link{performance}} object}

\item{bylabel}{(bool) should a different facet be plotted for each
level of \code{x$label}? (typically this is CV repetitions)}

\item{log}{(bool) should x and y axes be log(1 + x) transformed
(useful for poisson)}

\item{jitter}{(numeric) if nonzero then the requested amount of
jitter will be applied to x and y values.}

\item{plot}{(bool) produce the plot? or just return the plot and data.}
}
\value{
a list containing the plot and source data.
}
\description{
Plot the prediction error of a dCVnet performance object.
}
