% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_referencemodels.R
\name{print.reflogreg}
\alias{print.reflogreg}
\title{Simple print function for \code{\link{reflogreg}} objects.}
\usage{
\method{print}{reflogreg}(x, ...)
}
\description{
Simple print function for \code{\link{reflogreg}} objects.
}
