% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_classperformance.R
\name{report_classperformance_summary}
\alias{report_classperformance_summary}
\title{report_classperformance_summary}
\usage{
report_classperformance_summary(dCVnet_object)
}
\arguments{
\item{dCVnet_object}{result from a call to \code{\link{dCVnet}}}
}
\value{
a data.frame of summarised and raw performance statistics.
}
\description{
extracts classperformance from a dCVnet object
calculates classification statistics and
provides a summary of
\code{\link[base]{mean}},
\code{\link[stats]{sd}},
\code{\link[base:Extremes]{min}}, and \code{\link[base:Extremes]{max}}
}
