% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_referencemodels.R
\name{report_reference_classperformance_summary}
\alias{report_reference_classperformance_summary}
\title{report_reference_classperformance_summary}
\usage{
report_reference_classperformance_summary(refobj)
}
\arguments{
\item{refobj}{a set of reference models provided by
\code{\link{dCVnet_refmodels}}}
}
\value{
a data.frame of summarised and raw performance statistics.
}
\description{
Returns a report of classperformance summary information for a set of dCVnet
reference models.
}
