% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{selected_hyperparameters}
\alias{selected_hyperparameters}
\title{selected_hyperparameters}
\usage{
selected_hyperparameters(object, what = c("summary", "data", "both"), ...)
}
\arguments{
\item{object}{a dCVnet object}

\item{what}{desired returns:
\itemize{
\item{\code{"summary"} - alpha/lambda summary over all outer folds/reps.}
\item{\code{"data"} - return underlying data.}
\item{\code{"both"} - both of the above in a list.}
}}

\item{...}{" "}
}
\description{
What hyperparameters were chosen for the outer loop of a dCVnet object.
}
