% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_classperformance.R
\name{summary.classperformance}
\alias{summary.classperformance}
\title{summary.classperformance}
\usage{
\method{summary}{classperformance}(object, label = NA, pvprevalence = "observed", ...)
}
\arguments{
\item{object}{a \code{\link{classperformance}} object.}

\item{label}{a label can be assigned here.
(Warning - setting a length 1 vector will concatenate multiple reps.)}

\item{pvprevalence}{argument for adjustment of PPV/NPV calculation.
either "observed", or number \code{[0,1]}.}

\item{...}{additional arguments (ignored)}
}
\description{
Calculates classification performance table and
two-class classification metrics for a
\code{\link{classperformance}} object.
}
