% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{summary.dCVnet}
\alias{summary.dCVnet}
\title{summary.dCVnet}
\usage{
\method{summary}{dCVnet}(object, ...)
}
\arguments{
\item{object}{a a \code{\link{dCVnet}} object.}

\item{...}{NULL}
}
\value{
a dataframe of tuning results for each outer fold/rep.
[Note: this may change in future]
}
\description{
a summary of key options and results for a
\code{\link{dCVnet}} object.
}
\details{
Prints the following sections: \itemize{
\item{Input descriptives}
\item{Outer Loop classification performance}
\item{Inner Loop model stability}
\item{Inner Loop model performance}
\item{Production model performance (fit to \emph{all} data)}
}
The outer loop tests generalisation of the model.
}
