% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{summary.multialpha.repeated.cv.glmnet}
\alias{summary.multialpha.repeated.cv.glmnet}
\title{summary.multialpha.repeated.cv.glmnet}
\usage{
\method{summary}{multialpha.repeated.cv.glmnet}(object, print = T, ...)
}
\arguments{
\item{object}{a a \code{\link{multialpha.repeated.cv.glmnet}} object.}

\item{print}{if FALSE silently returns the summary results table.}

\item{...}{NULL}
}
\description{
a summary of key options and results for a
\code{\link{multialpha.repeated.cv.glmnet}} object.
}
