% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_innerloop.R
\name{summary.repeated.cv.glmnet}
\alias{summary.repeated.cv.glmnet}
\title{summary.repeated.cv.glmnet}
\usage{
\method{summary}{repeated.cv.glmnet}(object, ...)
}
\arguments{
\item{object}{a a \code{\link{repeated.cv.glmnet}} object.}

\item{...}{NULL}
}
\description{
a summary of key options and results for a \code{\link{repeated.cv.glmnet}}
object.
}
