% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{tidy_predict.glmnet}
\alias{tidy_predict.glmnet}
\title{tidy_predict.glmnet}
\usage{
tidy_predict.glmnet(
  mod,
  newx,
  s,
  family,
  newy = NULL,
  newoffset = NULL,
  label = "",
  binomial_thresh = 0.5,
  ...
)
}
\arguments{
\item{mod}{a fitted glmnet object (alpha is determined by the object)}

\item{newx}{new values of x for which predictions are desired.}

\item{s}{the value of penalty parameter lambda at which predictions are
required.}

\item{family}{the glmnet model family
(this determines the format of the return)}

\item{newy}{outcome associated with newx. If provided these will be included
in the output (useful for subsequent performance assessment).}

\item{newoffset}{if an offset is used in the fit, then one must be supplied
for making predictions.}

\item{label}{an optional label (value is added in column "label")}

\item{binomial_thresh}{this allows non-default thresholds to
be used for classification. This is only relevant for binary
classification. E.g. for an imbalanced binary outcome
with 70:30 allocation, setting the decision threshold to 0.7
gives a better balance of sensitivity and specificity
without requiring threshold tuning (as in AUC optimal threshold).}

\item{...}{passed to \code{\link{predict.glmnet}}}
}
\value{
a \code{\link{data.frame}} containing column(s) for:
\itemize{
\item{'prediction' - result of
\code{predict.glmnet(.., type = "response")}.
The interpretation depends on the model family.
}
\item{'rowid' - the rownames of newx}
\item{'label' - a column of a single label used when merging predictions}
}
Optionally the data.frame will contain:
\itemize{
\item{'classification' - the predicted class (for nominal outcomes).}
\item{'reference' - the response being predicted (if newy specified).}
}
}
\description{
return a dataframe of glmnet predictions associated with outcomes (when these
are provided.)
}
