% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{tuning_plot_dCVnet}
\alias{tuning_plot_dCVnet}
\title{tuning_plot_dCVnet}
\usage{
tuning_plot_dCVnet(object, n.random = 0, plot = TRUE)
}
\arguments{
\item{object}{a \code{\link{dCVnet}} object}

\item{n.random}{select a random sample of k-fold reps to display.
0 = display all.}

\item{plot}{(bool) produce the plot? or just return the plot and data}
}
\value{
a list containing the plot and a data.frame used to plot
(full data.frame is returned, i.e. ignores n.random)
}
\description{
Display the tuning curves for the repeated k-folds of the outerloop of a
dCVnet.
}
\details{
requires \pkg{ggplot2}.
}
