% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{variable_importance}
\alias{variable_importance}
\title{variable_importance}
\usage{
variable_importance(x, scale = FALSE, percentage = FALSE)
}
\arguments{
\item{x}{a dCVnet object}

\item{scale}{Boolean. Should the return values be scaled so the most
important value is 1?}

\item{percentage}{Boolean. Should the return values be scaled so the most
important value is 100?}
}
\value{
a data.frame of variable names "Predictor"
and variable importance "varImp".
}
\description{
Variable importance for dCVnet/glmnet models does not require permutation
methods, because coefficients are directly interpretable.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This VI function follow caret's example (see \code{\link[caret]{varImp}}
function) and simply returns the absolute values of the coefficients.

As variable importance is inferential this is done for the tuned
"production" model rather than the cross-validated outer-loop.
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[caret]{varImp}}
}
