% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_classperformance.R
\name{classperformance}
\alias{classperformance}
\title{classperformance}
\usage{
classperformance(x, ...)
}
\arguments{
\item{x}{an object from which binary class performance can be extracted.}

\item{...}{arguments to pass on}
}
\value{
a classperformance object, is a dataframe (or list of dataframes)
with the following columns:
\itemize{
\item{reference - the known 'true' class of the observation}
\item{classification - the predicted class assigned by the model}
\item{probability - the probability of the non-positive class assigned
by the model (used for AUROC)}
\item{label - a grouping variable when predictions come from more than
one source, e.g. multiple reps}
}
}
\description{
extracts a standardised classification performance table for a model.
}
