% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_classperformance.R
\name{classperformance}
\alias{classperformance}
\alias{classperformance.list}
\alias{classperformance.classperformance}
\alias{classperformance.dCVnet}
\alias{classperformance.glm}
\alias{classperformance.glmlist}
\title{classperformance}
\usage{
classperformance(x, ...)

\method{classperformance}{list}(x, ...)

\method{classperformance}{classperformance}(x, as.data.frame = TRUE, ...)

\method{classperformance}{dCVnet}(x, as.data.frame = TRUE, ...)

\method{classperformance}{glm}(x, as.data.frame = TRUE,
  label = deparse(substitute(x)), threshold = 0.5, newdata = NULL,
  ...)

\method{classperformance}{glmlist}(x, as.data.frame = TRUE, ...)
}
\arguments{
\item{x}{an object from which binary class performance can be extracted.}

\item{...}{arguments to pass on}

\item{as.data.frame}{return a data.frame instead of a list of
\code{\link{classperformance}} objects.}

\item{label}{specify a label for the output}

\item{threshold}{logistic regression only - use a threshold other than 0.5.}

\item{newdata}{evaluate performance in new data}
}
\value{
a classperformance object, is a dataframe (or list of dataframes)
with the following columns:
\itemize{
\item{reference - the known 'true' class of the observation}
\item{prediction - the model prediction for a case.
for dCVnet this is the result of predict(model, type = "response")
With "binary" response predictions are the predicted probability
of the non-reference class (used for AUROC)}
\item{classification - for binomial and multinomial models this is
the predicted class assigned by the model}
\item{label - a grouping variable when predictions come from more than
one source, e.g. multiple reps}
}
}
\description{
extracts a standardised classification performance table for a model.

For glm objects wraps \link{predict.glm} if newdata is specified.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: classperformance for \code{\link{list}} object
(treats it as a \code{\link{dCVnet}} object)

\item \code{classperformance}: returns self (allows list/dataframe conversion)

\item \code{dCVnet}: classperformance for \code{\link{dCVnet}} object

\item \code{glm}: classperformance for \code{\link[stats]{glm}}
object

\item \code{glmlist}: classperformance for glmlist from
\code{\link{reflogreg}} object
}}

