% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{coef.dCVnet}
\alias{coef.dCVnet}
\title{coef.dCVnet}
\usage{
\method{coef}{dCVnet}(object, type = "all", ...)
}
\arguments{
\item{object}{a dCVnet object}

\item{type}{how to return coefficients.
\itemize{
\item{\code{"production"} - the "production" model coefficients}
\item{\code{"all"} - return separate coefficients
for each fold & rep of CV}
\item{\code{"mean"} - return mean over \code{"all"}.}
\item{\code{"median"} - return median over \code{"all"}.}
\item{\code{"byrep"} - return separate coefficients for each rep of CV
(take mean average over folds).}
\item{\code{byrep_mean} - mean over \code{"byrep"}}
\item{\code{byrep_mean} - median over \code{"byrep"}}
}}

\item{...}{" "}
}
\value{
a data.frame of coefficient values (see type argument) containing
columns: Predictor and Coef (as well as Rep and fold if applicable)
}
\description{
Coefficients from a dCVnet object. If type is "production" this gives the
coefficients from the production model fit to all data. Otherwise
coefficients are returned from the outerloop of the cross-validation.
}
\seealso{
coefficients_summary
}
