% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{coef.dCVnet}
\alias{coef.dCVnet}
\title{coef.dCVnet}
\usage{
\method{coef}{dCVnet}(object, type = "all", ...)
}
\arguments{
\item{object}{a dCVnet object}

\item{type}{how to return coefficients.
\itemize{
\item{\code{"all"} - return separate coefficients for each rep/fold.}
\item{\code{"rep"} - return separate coefficients for each rep
(mean average over folds).}
\item{\code{"mean"} - return mean of \code{"rep"}.}
\item{\code{"median"} - return median of \code{"rep"}.}
}}

\item{...}{" "}
}
\description{
Coefficients from a dCVnet object.
}
