% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{coefficients_summary}
\alias{coefficients_summary}
\title{coefficients_summary}
\usage{
coefficients_summary(object, ...)
}
\arguments{
\item{object}{a dCVnet object}

\item{...}{" "}
}
\value{
a data.frame containing
}
\description{
Extract the coefficients in the production model (fit to complete data)
and provide descriptives for the coefficients in the outerloop
cross-validation
}
\seealso{
coef.dCVnet
}
