% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet.R
\docType{package}
\name{dCVnet-package}
\alias{dCVnet-package}
\title{dCVnet: Doubly Cross-Validated Elastic-Net Regularised (Generalised)
Linear Models}
\description{
dCVnet runs binary logistic regression with elastic-net regulatisation.
Optimal hyperparameters are selected by double (nested) cross-validation with
model performance evaluated in an independent outer loop.
}
\details{
The \code{lambda} and \code{alpha} hyperparameters of the elastic-net
allow models to
range from effectively unregularised, to heavily-regularised. Further a
mixture of two types of regularisation: L2 (ridge) and L1 (LASSO) are
possible. This regularisation produces dimensionality reduction and
variable selection in the predictors.

The values of the hyperparameters and thereby the amount and type of
regularisation is selected on the basis of what
hyperparameter combinations produce performance which generalises the best
in the inner cross-validation.

The fully nested cross-validation keeps the final performance estimates
'honest' by reducing optimism bias.
}
