% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet.R
\docType{package}
\name{dCVnet-package}
\alias{dCVnet-package}
\title{dCVnet: Doubly Cross-Validated Elastic-Net Regularised (Generalised)
Linear Models}
\description{
dCVnet fits and cross-validates regression with elastic-net regularisation.
Optimal hyperparameters (lambda & alpha) are selected by double (nested)
cross-validation with model performance evaluated in an independent outer
cross-validation loop.
}
\details{
The \code{lambda} and \code{alpha} hyperparameters of the elastic-net
allow models to freely
range from effectively unregularised, to heavily-regularised (via lambda).
With a balance of two types of regularisation: L2 (ridge) and L1 (LASSO)
(alpha). This regularisation produces dimensionality reduction and
variable selection in the predictors.

The values of the lambda and alpha hyperparameters (and so the amount
and type of regularisation) are selected on the basis of cross-validated
model performance - i.e. a data-driven approach.

Using a fully nested cross-validation (instead of reporting the CV
performance of the selected hyperparameters) keeps the cross-validated
performance estimates 'honest' by reducing optimism bias related to
hyperparameter tuning.
}
