% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_referencemodels.R
\name{dCVnet_refmodels}
\alias{dCVnet_refmodels}
\title{dCVnet_refmodels}
\usage{
dCVnet_refmodels(object)
}
\arguments{
\item{object}{a dCVnet object}
}
\value{
a list of \code{glm} and \code{univariate} models.
}
\description{
calculate some reference models to help interpret dCVnet performance
models calculated are:
\itemize{
\item{a logistic regression using all variables (if n>>k), otherwise
a logistic regression on the top n/5 principal compoments}
\item{a series of logistic regressions, one for each column in the
design matrix - a mass univariate approach}
}
}
