% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_roc.R
\name{extract_rocdata}
\alias{extract_rocdata}
\title{extract_rocdata}
\usage{
extract_rocdata(classperformance, ROCR.invertprob = T)
}
\arguments{
\item{classperformance}{a \code{\link{classperformance}} object.}

\item{ROCR.invertprob}{boolian. Should class probabilities be inverted?
This will depend on what your 'positive' class is.
The internal \link[ROCR]{prediction} uses opposing convention to the
rest of this package, so normally this should be TRUE.
\itemize{
\item{ROCR}{last level is 'positive'}
\item{typical}{first level is 'positive'}
}
If the ROC curve appears under the diagonal then toggle this option.}
}
\value{
a data.frame object of class "rocdata" which can be plotted. Contents:
\itemize{
\item{\code{Sens} : Sensitivity}
\item{\code{InvSpec} : 1 - Specificity}
\item{\code{alpha} : threshold}
\item{\code{run} : label from \code{\link{classperformance}}}
}
}
\description{
Function converts a classperformance data.frame to the values required for
calculating a ROC curve. It does this by calculating sensitivity and
specificity at incrementing class probability thresholds.
}
\details{
Uses \pkg{ROCR} functions: \code{\link[ROCR]{prediction}}, and
\code{\link[ROCR]{performance}}
}
\seealso{
\code{\link{plot.rocdata}}
}
