% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_performance.R
\name{perf_nomial}
\alias{perf_nomial}
\title{InternalPerformanceSummaryFunctions}
\usage{
perf_nomial(object, short = FALSE, pvprevalence = "observed")
}
\arguments{
\item{object}{a \code{\link{performance}} object.}

\item{short}{(bool) return a core set of performance measures.}

\item{pvprevalence}{argument for adjustment of PPV/NPV calculation.
either "observed", or number \code{[0,1]}.}
}
\description{
Internal summary functions for performance objects from
different model families.
}
\details{
Used for both binomial and multinomial families
}
