% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_performance.R
\name{InternalPerformanceSummaryFunctions}
\alias{InternalPerformanceSummaryFunctions}
\alias{perf_nomial}
\alias{perf_cont}
\alias{perf_cox}
\title{Internal summary functions for performance objects from
different model families.}
\usage{
perf_nomial(object, short = FALSE, pvprevalence = "observed")

perf_cont(object, short = FALSE, pvprevalence = "observed")

perf_cox(object, short = FALSE, pvprevalence = "observed")
}
\description{
Internal summary functions for performance objects from
different model families.
}
\section{Functions}{
\itemize{
\item \code{perf_nomial}: Used for binomial and multinomial families

\item \code{perf_cont}: Used for gaussian and poisson families

\item \code{perf_cox}: Used for cox models
}}

