% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_performance.R
\name{InternalPerformanceSummaryFunctions}
\alias{InternalPerformanceSummaryFunctions}
\alias{perf_nomial}
\alias{perf_cont}
\alias{perf_cox}
\alias{perf_mgaussian}
\title{Internal summary functions for performance objects from
different model families.}
\usage{
perf_nomial(object, short = FALSE, pvprevalence = "observed")

perf_cont(object, short = FALSE, pvprevalence = "observed")

perf_cox(object, short = FALSE, pvprevalence = "observed")

perf_mgaussian(object, short = FALSE, pvprevalence = "observed")
}
\arguments{
\item{object}{a \code{\link{performance}} object.}

\item{short}{(bool) return a core set of performance measures.}

\item{pvprevalence}{argument for adjustment of PPV/NPV calculation for
binomial or multinomial families. Either "observed" to use the observed
prevalence, or a number \code{[0,1]} (for binomial),
or a vector of length n_categories (for multinomial) containing numerics
in \code{[0,1]}.}
}
\description{
Internal summary functions for performance objects from
different model families.
}
\section{Functions}{
\itemize{
\item \code{perf_nomial}: Used for binomial and multinomial families

\item \code{perf_cont}: Used for gaussian and poisson families

\item \code{perf_cox}: Used for cox models

\item \code{perf_mgaussian}: Used for multivariate gaussian models
}}

