% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{parse_dCVnet_input}
\alias{parse_dCVnet_input}
\title{parse_dCVnet_input}
\usage{
parse_dCVnet_input(data, y, family, f = "~.", positive = 1,
  offset = NULL, yname = "y", passNA = FALSE)
}
\arguments{
\item{data}{a data.frame containing all terms in f.}

\item{y}{the outcome (can be numeric vector,
a factor (for binomial / multinomial) or a matrix for cox/mgaussian)}

\item{family}{the model family (see \code{\link{glmnet}})}

\item{f}{a one sided formula.
The RHS must refer to columns in \code{data} and may include
interactions, transformations or expansions (like \code{\link{poly}}, or
\code{\link{log}}).
The formula \emph{must} include an intercept.}

\item{positive}{What level of the outcome is a 'positive' result
(in the sense of a diagnostic test).
Can be a numeric indicating which of \code{levels(y)} to use
(i.e. 1 | 2). Alternatively a character specifying the exact level
(e.g. \code{"patient"}).}

\item{offset}{optional model offset (see \code{\link{glmnet}})}

\item{yname}{an optional label for the outcome / y variable.}

\item{passNA}{should NA values be excluded (FALSE) or passed through (TRUE)?}
}
\value{
\itemize{
\item{ \code{y} - outcome factor
with level ordered according to \code{positive}}
\item{ \code{x_mat} - predictor matrix
including expansions, interaction terms specified in \code{f}}
}
}
\description{
Collate a formula and dataset into a standardised object ready for dCVnet
}
\details{
The outcome is coerced to a binary factor which is releveled (if necessary)
such that the test=positive level comes first.
}
