% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{plot.multialpha.repeated.cv.glmnet}
\alias{plot.multialpha.repeated.cv.glmnet}
\title{plot.multialpha.repeated.cv.glmnet}
\usage{
\method{plot}{multialpha.repeated.cv.glmnet}(x, xvar = c("lambda", "s"), errorbars = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{multialpha.repeated.cv.glmnet}} object.}

\item{xvar}{what to plot on x-axis. Either the log10 lambda value, or the
proportion of the lambda sequence (for that alpha).}

\item{errorbars}{boolean. Add errorbars to the plot?}

\item{plot}{(bool) produce the plot? or just return the plot and data}

\item{...}{NULL}
}
\description{
Plot averaged and best innerloop model performance highlighting
the selected alpha for a
\code{\link{multialpha.repeated.cv.glmnet}}
}
