% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{plot_outerloop_coefs}
\alias{plot_outerloop_coefs}
\title{plot_outerloop_coefs}
\usage{
plot_outerloop_coefs(
  object,
  type = "rep",
  ordered = FALSE,
  abs = FALSE,
  intercept = FALSE,
  prod = TRUE,
  prod_col = "red",
  prod_shape = 24,
  panel_scaling = c("free", "fixed"),
  plot = TRUE
)
}
\arguments{
\item{object}{a \code{\link{dCVnet}} object}

\item{type}{Use "all" to inspect variability
over cross-validation folds and reps. Use "production" for only the final
model. See \code{\link{coef.dCVnet}} for full list of options.}

\item{ordered}{sort predictors by size?}

\item{abs}{plot absolute values?}

\item{intercept}{include the value of the intercept coefficient in the plot?}

\item{prod}{add the production model coefficients as an overlay?
(note: this is not in the data returned by this function,
but can be accessed with \code{coef(object, type = "production")})}

\item{prod_col}{colour for production model coefficients}

\item{prod_shape}{shape for production model coefficients}

\item{panel_scaling}{for multi-outcome coefficients (mgaussian, multinomial).
Should y-axes be independent, or same over all panels.}

\item{plot}{(bool) should the plot also be rendered (\code{TRUE})?,
or just returned as a R object (\code{FALSE}).}
}
\value{
a list containing the plot and a data.frame used to plot
(full data.frame is returned, i.e. ignores n.random)
}
\description{
Plot showing standardised betas for outer-loop model coefficients.
Because each fold/repetition of the outerloop can have
completely different amounts and types of regularisation
this should be interpreted with caution.
}
\details{
Warning: do not use these plots to select a subset of variables and
re-run dCVnet.
These coefficients are based on the complete dataset
and using the output of dCVnet to select variables will produce
optimism in cross-validated estimates of performance.
}
