% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{plot_outerloop_coefs}
\alias{plot_outerloop_coefs}
\title{plot_outerloop_coefs}
\usage{
plot_outerloop_coefs(object)
}
\arguments{
\item{object}{a dCVnet object}
}
\description{
Plot showing standardised betas for model coefficients.
coefficients are first mean-averaged over the k-folds,
then displayed per-repetition.
}
