% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{plot_outerloop_coefs}
\alias{plot_outerloop_coefs}
\title{plot_outerloop_coefs}
\usage{
plot_outerloop_coefs(
  object,
  type = "rep",
  ordered = FALSE,
  abs = FALSE,
  intercept = FALSE,
  final = TRUE,
  final_col = "red",
  final_shape = 24
)
}
\arguments{
\item{object}{a \code{\link{dCVnet}} object}

\item{type}{How to display coefficients.
passed to \code{\link{coef.dCVnet}} .
\itemize{
\item{\code{"all"} - boxplot of coefficients for each rep/fold.}
\item{\code{"rep"} - boxplot of mean coefficients for each rep
(mean average over folds).}
\item{\code{"mean"} - dotplot of the mean of \code{"rep"}.}
\item{\code{"median"} - dotplot of the median of \code{"rep"}.}
}}

\item{ordered}{sort predictors by size?}

\item{abs}{plot absolute values?}

\item{intercept}{include the value of the intercept coefficient in the plot?}

\item{final}{add the final model coefficients as an overlay?
(note: this is not in the return data, but can be accessed with
coef(object$final$model))}

\item{final_col}{colour for final model coefficients}

\item{final_shape}{shape for final model coefficients}
}
\description{
Plot showing standardised betas for outer-loop model coefficients.
Because each fold/repetition of the outerloop can have
completely different amounts and types of regularisation
this should be interpreted with caution.
}
\details{
Warning: do not use these plots to select a subset of variables and
re-run dCVnet.
These coefficients are based on the complete dataset
and using the output of dCVnet to select variables will produce
optimism in cross-validated estimates of performance.
}
