% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{repeated.cv.glmnet}
\alias{repeated.cv.glmnet}
\title{repeated.cv.glmnet}
\usage{
repeated.cv.glmnet(folds, lambdas, alpha, x, y, ..., debug = F)
}
\arguments{
\item{folds}{This is a list where each element is an integer vector
of length \emph{n_cases}. The integer for each case labels it as belonging
to a fold \emph{1:n_folds}. This argument implicitly sets the number of repeats
and the k in repeated k-fold cv.}

\item{lambdas}{use a fixed, user supplied lambda sequence (descending)
see \code{\link[glmnet]{glmnet}}}

\item{alpha}{The elasticnet mixing parameter, with
    \eqn{0\le\alpha\le 1}. The penalty is defined
    as \deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1}
    is the lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{x}{\code{x} matrix as in \code{glmnet}.}

\item{y}{response \code{y} as in \code{glmnet}.}

\item{...}{arguments passed to \code{\link[glmnet]{cv.glmnet}}}

\item{debug}{if TRUE return models and unaveraged results (default: FALSE).}
}
\value{
a data.frame object of class \code{\link{repeated.cv.glmnet}}
containting averaged metrics. Has the following columns:
\itemize{
\item{lambda - the specified lambda value}
\item{cvm - the averaged performance metric}
\item{cvsd - the averaged sd of performance metric}
\item{cvup - the averaged cvm + cvsd}
\item{cvlo - the averaged cvm - cvsd}
\item{nzero - the number of nonzero predictors}
\item{lambda.min - logical indicating best performing lambda
(on average)}
}
}
\description{
Repeatedly runs a \code{\link[glmnet]{cv.glmnet}} and returns averaged
results. \emph{This is intended to be a dCVnet internal function}.
}
