% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_innerloop.R
\name{repeated.cv.glmnet}
\alias{repeated.cv.glmnet}
\title{repeated.cv.glmnet}
\usage{
repeated.cv.glmnet(
  x,
  y,
  folds = NULL,
  lambda = NULL,
  alpha = 1,
  nfolds = NULL,
  nreps = NULL,
  family = c("gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"),
  ...,
  debug = FALSE
)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation
vector. Can be in sparse matrix format (inherit from class
\code{"sparseMatrix"} as in package \code{Matrix})}

\item{y}{response variable. Quantitative for \code{family="gaussian"}, or
\code{family="poisson"} (non-negative counts). For \code{family="binomial"}
should be either a factor with two levels, or a two-column matrix of counts
or proportions (the second column is treated as the target class; for a
factor, the last level in alphabetical order is the target class). For
\code{family="multinomial"}, can be a \code{nc>=2} level factor, or a matrix
with \code{nc} columns of counts or proportions. For either
\code{"binomial"} or \code{"multinomial"}, if \code{y} is presented as a
vector, it will be coerced into a factor. For \code{family="cox"}, preferably
a \code{Surv} object from the survival package: see Details section for
more information. For \code{family="mgaussian"}, \code{y} is a matrix
of quantitative responses.}

\item{folds}{This is a list where each element is an integer vector
of length \emph{n_cases}. The integer for each case labels it as belonging
to a fold \emph{1:n_folds}. This argument overrides the number of repeats
and the k in repeated k-fold cv.}

\item{lambda}{use a fixed, user supplied lambda sequence (descending)
see \code{\link[glmnet]{glmnet}}}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0\le\alpha\le 1}.
The penalty is defined as
\deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} is the
lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{nfolds}{if folds are not specified, how many folds should be used in
cross-validation? The default (nfolds=NULL) uses 10-fold.}

\item{nreps}{if folds are not specified, how many times to repeat k-fold
cross-validation? The default (nreps=NULL) uses 5 repeats.}

\item{family}{Either a character string representing
one of the built-in families, or else a \code{glm()} family object. For more
information, see Details section below or the documentation for response
type (above).}

\item{...}{arguments passed to \code{\link[glmnet]{cv.glmnet}}}

\item{debug}{return an unmerged list.}
}
\value{
a \code{\link{cv.glmnet}} object with cv performance averaged.
}
\description{
Repeatedly runs a \code{\link[glmnet]{cv.glmnet}} and returns averaged
results. \emph{This is intended as a dCVnet internal function}.
}
\details{
The code will run for any glmnet family, but folds & lambdas must be
correctly specified.
}
\examples{
\dontrun{
data("CoxExample", package = "glmnet") # x and y
mod <- repeated.cv.glmnet(x = CoxExample$x, y = CoxExample$y, family = "cox")
}
}
\seealso{
\code{\link{amalgamate_cv.glmnet}}
}
