% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_performance.R
\name{report_performance_summary}
\alias{report_performance_summary}
\title{report_performance_summary}
\usage{
report_performance_summary(dCVnet_object, pvprevalence = "observed")
}
\arguments{
\item{dCVnet_object}{result from a call to \code{\link{dCVnet}}}

\item{pvprevalence}{allows calculation of PPV/NPV at different prevalences.
set to "observed" to use the prevalence of the data.
Note: does not affect the presented prevalence value in the table.}
}
\value{
a data.frame of summarised and raw performance statistics.
}
\description{
extracts performance from a dCVnet object
calculates classification statistics and
provides a summary of
\code{\link[base]{mean}},
\code{\link[stats]{sd}},
\code{\link[base:Extremes]{min}}, and \code{\link[base:Extremes]{max}}
}
