% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_classperformance.R
\name{summary.classperformance}
\alias{summary.classperformance}
\title{summary.classperformance}
\usage{
\method{summary}{classperformance}(object, label = NA, ...)
}
\arguments{
\item{object}{a \code{\link{classperformance}} object.}

\item{label}{a label can be assigned here.}

\item{...}{additional arguments (not currently used)}
}
\description{
Calculates performance table and two-class classification metrics for a
\code{\link{classperformance}} object.
}
