% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_main.R
\name{summary.dCVnet}
\alias{summary.dCVnet}
\title{summary.dCVnet}
\usage{
\method{summary}{dCVnet}(object, ...)
}
\arguments{
\item{object}{a a \code{\link{dCVnet}} object.}

\item{...}{NULL}
}
\description{
a summary of key options and results for a
\code{\link{dCVnet}} object.
\emph{this function is not completely functional}
}
