% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_performance.R
\name{summary.performance}
\alias{summary.performance}
\title{summary.performance}
\usage{
\method{summary}{performance}(
  object,
  label = NA,
  short = FALSE,
  pvprevalence = "observed",
  somersD = FALSE,
  ...
)
}
\arguments{
\item{object}{a \code{\link{performance}} object.}

\item{label}{a label can be assigned here.
(Warning - setting a length 1 vector will concatenate multiple reps.)}

\item{short}{(bool) return a core set of performance measures.}

\item{pvprevalence}{argument for adjustment of PPV/NPV calculation for
binomial or multinomial families. Either "observed" to use the observed
prevalence, or a number \code{[0,1]} (for binomial),
or a vector of length n_categories (for multinomial) containing numerics
in \code{[0,1]}.}

\item{somersD}{(bool) calculate the computationally expensive Somers' D for
certain families (gaussian, poisson)}

\item{...}{additional arguments (ignored)}
}
\description{
Calculates classification performance table and
two-class classification metrics for a
\code{\link{performance}} object.
}
