% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{tidy_coef.glmnet}
\alias{tidy_coef.glmnet}
\title{tidy_coef.glmnet}
\usage{
tidy_coef.glmnet(mod, s, label = "coef")
}
\arguments{
\item{mod}{a fitted glmnet object (alpha is determined by the object)}

\item{s}{the value of penalty parameter lambda at which predictions are
required.}

\item{label}{an optional label (value is added in column "label")}
}
\description{
return a dataframe of glmnet coefficients
}
