% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_utilities.R
\name{tidy_coef.glmnet}
\alias{tidy_coef.glmnet}
\title{tidy_coef.glmnet}
\usage{
tidy_coef.glmnet(
  mod,
  newx,
  newy,
  s,
  family,
  label = "",
  newoffset = NULL,
  binomial_thresh = 0.5
)
}
\arguments{
\item{mod}{a fitted glmnet object (alpha is determined by the object)}

\item{newx}{new values of x for which predictions are desired.}

\item{newy}{outcome associated with newx. If provided these will be included
in the output (useful for subsequent performance assessment).}

\item{s}{the value of penalty parameter lambda at which predictions are
required.}

\item{family}{the glmnet model family}

\item{label}{an optional label (value is added in column "label")}

\item{newoffset}{if an offset is used in the fit, then one must be supplied
for making predictions.}

\item{binomial_thresh}{this allows non-default thresholds to
be used for classification. This is only relevant for binary
classification. E.g. for an imbalanced binary outcome
with 70:30 allocation, setting the decision threshold to 0.7
gives a better balance of sensitivity and specificity
without requiring threshold tuning (as in AUC optimal threshold).}
}
\description{
return a dataframe of glmnet predictions associated with outcomes (when these
are provided.)
}
