% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{tuning_plot_dCVnet}
\alias{tuning_plot_dCVnet}
\title{tuning_plot_dCVnet}
\usage{
tuning_plot_dCVnet(object, n.random = 0)
}
\arguments{
\item{object}{a \code{\link{dCVnet}} object}

\item{n.random}{select a random sample of k-fold reps to display.
0 = display all.}
}
\value{
a data.frame containing the full dataset used to plot
(ignores n.random)
}
\description{
Display the tuning curves for the repeated k-folds of the outerloop of a
dCVnet.
}
\details{
requires \pkg{ggplot2}.
}
