% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCVnet_plotting.R
\name{tuning_plot_dCVnet}
\alias{tuning_plot_dCVnet}
\title{tuning_plot_dCVnet}
\usage{
tuning_plot_dCVnet(object)
}
\arguments{
\item{object}{a \code{\link{dCVnet}} object}
}
\value{
a ROC plot, as above.
}
\description{
Display the tuning curves for the repeated k-folds of the outerloop of a
dCVnet.
}
\details{
requires \pkg{ggplot2}.
}
